+++ frontend/app/src/components/index.tsx
@@ -1,0 +1,7 @@
+import Navbar from "./layout/Navbar";
+import Sidebar from "./layout/Sidebar";
+
+export {
+    Navbar,
+    Sidebar
+}
