+++ frontend/app/src/main.tsx
@@ -1,17 +1,13 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import { BrowserRouter, Routes, Route } from 'react-router'
+import { BrowserRouter } from 'react-router'
 import './index.css'
 import App from './App.tsx'
-import {DevAMPL} from './pages'
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
     <BrowserRouter>
-      <Routes>
-        <Route path='/' element={<App/>}></Route>
-        <Route path='/dev' element={<DevAMPL/>}></Route>
-      </Routes>
+      <App/>
     </BrowserRouter>
   </StrictMode>,
 )
