+++ frontend/app/src/pages/index.tsx
@@ -1,5 +1,7 @@
 import DevAMPL from "./DevAMPL";
+import Home from "./Home";
 
 export {
-    DevAMPL
+    DevAMPL,
+    Home
 }
